Rename ComponentInspector to TreeInspector
